refactor: move functions in helpers file.
